Add files via upload
This website will help you perform Ruqyah.
Ruqyah is a form of Islamic spiritual healing that involves reciting verses from the Quran and supplications to seek protection and cure from illnesses. It is believed to be a powerful tool that can heal both the body and mind. In this blog post, we will discuss the benefits of Ruqyah and how it can help individuals struggling with physical and mental health issues.
Protection from Evil:
One of the core beliefs of Ruqyah is that it can provide protection from evil and negative energies. Reciting Quranic verses can act as a shield against black magic, envy, and other harmful energies that can harm an individual's physical and mental health.